Move Cron class to helper file to consolidate code as per feedback
